refactor: 상품 카드 분리
