+++ src/refactoring/components/index.ts
@@ -1,2 +1,3 @@
 export * from './PageLayout';
 export * from './Navigation';
+export * from './ProductCard';
