+++ src/refactoring/hooks/useCart.ts
@@ -1,11 +1,7 @@
 // useCart.ts
 import { useState } from 'react';
 import { CartItem, Coupon, Product } from '../../types';
-import {
-  addProductToCart,
-  calculateCartTotal,
-  updateCartItemQuantity,
-} from '../models/cart';
+import { addProductToCart, updateCartItemQuantity } from '../models/cart';
 import { useLocalStorage } from './useLocalStorage';
 
 export const useCart = () => {
